refactor : Repository JPA와 Elasticsearch 분리
