+++ projecthub/src/main/java/com/project/hub/service/impl/UserLikeService.java
@@ -12,10 +12,10 @@ import com.project.hub.model.dto.request.likes.ProjectLikeRequest;
 import com.project.hub.model.dto.response.ResultResponse;
 import com.project.hub.model.type.LikeType;
 import com.project.hub.model.type.ResultCode;
-import com.project.hub.repository.CommentsLikeRepository;
-import com.project.hub.repository.CommentsRepository;
-import com.project.hub.repository.ProjectRepository;
-import com.project.hub.repository.ProjectsLikeRepository;
+import com.project.hub.repository.jpa.CommentsLikeRepository;
+import com.project.hub.repository.jpa.CommentsRepository;
+import com.project.hub.repository.jpa.ProjectRepository;
+import com.project.hub.repository.jpa.ProjectsLikeRepository;
 import com.project.hub.service.LikeService;
 import jakarta.transaction.Transactional;
 import java.util.Set;
